Analytics: Add support for writing log files

+++ EVTCAnalytics/Parsed/ParsedAgent.cs
@@ -6,14 +6,14 @@ namespace GW2Scratch.EVTCAnalytics.Parsed
 		public string Name { get; }
 		public uint Prof { get; }
 		public uint IsElite { get; }
-		public int Toughness { get; }
-		public int Concentration { get; }
-		public int Healing { get; }
-		public int Condition { get; }
-		public int HitboxWidth { get; }
-		public int HitboxHeight { get; }
+		public short Toughness { get; }
+		public short Concentration { get; }
+		public short Healing { get; }
+		public short Condition { get; }
+		public short HitboxWidth { get; }
+		public short HitboxHeight { get; }
 
-		public ParsedAgent(ulong address, string name, uint prof, uint isElite, int toughness, int concentration, int healing, int condition, int hitboxWidth, int hitboxHeight)
+		public ParsedAgent(ulong address, string name, uint prof, uint isElite, short toughness, short concentration, short healing, short condition, short hitboxWidth, short hitboxHeight)
 		{
 			Address = address;
 			Name = name;
